requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
